Restructured wr_logic_longrange
