--- src/wr_logic_ai/launch/fake_data.launch
@@ -1,3 +1,0 @@
-<launch>
-  <node name="fake_data" pkg="wr_logic_ai" type="inputFakeData.py" output="screen"/>
-</launch>
