+++ src/wr_logic_ai/launch/state_machine.launch
@@ -20,13 +20,13 @@
     if="$(eval env('WROVER_HW') == 'REAL')" />
 
   <!-- Launch files when simulating data -->
-  <node name="fake_data" pkg="wr_logic_ai" type="inputFakeData.py" output="screen"
+  <node name="fake_data" pkg="wr_logic_longrange" type="inputFakeData.py" output="screen"
     unless="$(eval env('WROVER_HW') == 'REAL')" />
 
   <node pkg="wr_logic_ai" name="wait_for_user_input" type="wait_for_user_input.py" output="screen" />
   <!-- <node pkg="wr_logic_ai" name="nav_state_machine_driver" type="state_machine.py" output="screen" /> -->
 
-  <include file="$(find wr_logic_ai)/launch/long_range.launch" />
+  <include file="$(find wr_logic_longrange)/launch/long_range.launch" />
 
   <!-- <include file="$(find wr_led_matrix)/launch/led_matrix.launch" /> -->
