+++ src/wr_logic_longrange/launch/fake_data.launch
@@ -1,0 +1,3 @@
+<launch>
+  <node name="fake_data" pkg="wr_logic_longrange" type="inputFakeData.py" output="screen"/>
+</launch>
